ca55: Fix support for clean build

+++ application_cpu/build_xenhypervisor.sh
@@ -49,53 +49,55 @@ SCRIPT_DIR=$(cd `dirname $0` && pwd)
 AOS_VERSION="v1.0.0"
 
 # Prepare working directory
-cd ${SCRIPT_DIR}
-rm -rf ./work
-mkdir -p ./work
-cd ./work
+if [[ ! -e "${SCRIPT_DIR}/work" || "$CLEAN_BUILD_FLAG" == "true" ]]; then
+    cd ${SCRIPT_DIR}
+    rm -rf ./work
+    mkdir -p ./work
+    cd ./work
 
-# Preprae yaml file
-cat ../aos-rcar-gen4.yaml ../aos-rcar-gen4-patch.yaml > ./aos-rcar-gen4-wb.yaml
+    # Preprae yaml file
+    cat ../aos-rcar-gen4.yaml ../aos-rcar-gen4-patch.yaml > ./aos-rcar-gen4-wb.yaml
 
-#############################################
-# START: Apply patch for meta-aos-rcar-gen4 #
-#############################################
-# Remove meta-aos-rcar-gen from repo list
-PARTERN='    - type: git
+    #############################################
+    # START: Apply patch for meta-aos-rcar-gen4 #
+    #############################################
+    # Remove meta-aos-rcar-gen from repo list
+    PARTERN='    - type: git
       url: "https://github.com/aoscloud/meta-aos-rcar-gen4.git"
       rev: "v1.0.0"'
-sed -i -z "s|${PARTERN//$'\n'/\\n}||" ./aos-rcar-gen4-wb.yaml
+    sed -i -z "s|${PARTERN//$'\n'/\\n}||" ./aos-rcar-gen4-wb.yaml
 
-# Prepare additional repo
-mkdir -p ./yocto
-git clone https://github.com/aoscloud/meta-aos-rcar-gen4 \
+    # Prepare additional repo
+    mkdir -p ./yocto
+    git clone https://github.com/aoscloud/meta-aos-rcar-gen4 \
     ./yocto/meta-aos-rcar-gen4
-cd ./yocto/meta-aos-rcar-gen4
-git checkout ${AOS_VERSION}
-cd ../../
+    cd ./yocto/meta-aos-rcar-gen4
+    git checkout ${AOS_VERSION}
+    cd ../../
 
-# Apply patch
-cd ./yocto/meta-aos-rcar-gen4
-git am ../../../patchset_aos/*
+    # Apply patch
+    cd ./yocto/meta-aos-rcar-gen4
+    git am ../../../patchset_aos/*
 
-# Patch for S4SK
-if [ "$1" == "s4sk" ]; then
-    git am ../../../patchset_s4sk/*
+    # Patch for S4SK
+    if [ "$1" == "s4sk" ]; then
+        git am ../../../patchset_s4sk/*
+    fi
+
+    cd ../../
+    #############################################
+    # END: Apply patch for meta-aos-rcar-gen4   #
+    #############################################
 fi
 
-cd ../../
-
-#############################################
-# END: Apply patch for meta-aos-rcar-gen4   #
-#############################################
-
+cd ${SCRIPT_DIR}/work
 moulin ./aos-rcar-gen4-wb.yaml \
     --TARGET_BOARD $1 \
     --USING_UFS_AS_STORAGE $USING_UFS
 ninja
 ninja image-full
-mv full.img $1.${BOOT_DEV}.full.img
-gzip $1.${BOOT_DEV}.full.img
+mv -f full.img $1.${BOOT_DEV}.full.img
+gzip -f $1.${BOOT_DEV}.full.img
 if [[ -e "${SCRIPT_DIR}/work/yocto/build-domd/tmp/deploy/sdk" ]]; then
     find ${SCRIPT_DIR}/work/yocto/build-domd/tmp/deploy/sdk/ -name *.sh | xargs cp -f -t ${SCRIPT_DIR}/work/
 fi
